Corrections possibles pour warnings avec Qt >= 5.9.5 et MSVC >= 2015
opencv_default_path.pri
    ==> Montée de version possible (testée)
pcl_default_path.pri
    ==> Montée de version possible (testée)
AMKgl\tools\opengl\ellipseglrenderer.cpp
    ==> Redefinition de 'M_xxx' via QtMath -> 'include' après QtMath, et utilisation du define _USE_MATH_DEFINES
AMKgl\tools\opengl\shapetovolume.cpp
    ==> Redefinition de 'M_xxx' via QtMath -> 'include' après QtMath, et utilisation du define _USE_MATH_DEFINES
ComputreeGui\view\DocumentView\GraphicsViews\3D\Camera\g3dcameracontroller.cpp
    ==> Redefinition de 'M_xxx' via QtMath -> 'include' après QtMath, et utilisation du define _USE_MATH_DEFINES
library\ctlibstructureaddon\ct_itemdrawable\tools\scanner\ct_thetaphishootingpattern.cpp
    ==> Redefinition de 'M_xxx' via QtMath -> 'include' après QtMath, et utilisation du define _USE_MATH_DEFINES
ComputreeGui\dm_guimanager.cpp
    ==> Argument de fonction non utilisé, sur des fonctions abstraites et/ou en cours de développement
ComputreeGui\tools\amkgl\dm_objectsvisitorfollowerbymodel.cpp
    ==> Argument de fonction non utilisé, sur des fonctions abstraites et/ou en cours de développement
ComputreeGui\view\DocumentView\GraphicsViews\3D\ggraphicsviewimp.cpp
    ==> Argument de fonction non utilisé, sur des fonctions abstraites et/ou en cours de développement
library\ctlibcore\cdm_hrscriptmanagerxml.cpp
    ==> Argument de fonction non utilisé, sur des fonctions abstraites et/ou en cours de développement
library\ctlibcore\cdm_hrscriptxmlreader.cpp
    ==> Argument de fonction non utilisé, sur des fonctions abstraites et/ou en cours de développement
library\ctlibmodels\ct_categories\abstract\ct_abstractcategory.cpp
    ==> Argument de fonction non utilisé, sur des fonctions abstraites et/ou en cours de développement
ComputreeGui\tools\attributes\worker\dm_attributesmanagerupdater.cpp
    ==> Déclaration de fonctions manquantes pour coller aux headers
ComputreeGui\view\DocumentView\GraphicsViews\PointsAttributes\gpointsattributesmanager.cpp
    ==> Déclaration de fonctions manquantes pour coller aux headers
ComputreeGui\view\StepResultTreeView\gtreestepcontextmenu.cpp
    ==> Passage de 'this' via la lambda contenant un Qt::connect
library\ctlibstructureaddon\ct_itemdrawable\ct_topfnodegroup.cpp
    ==> Conversion de type implicite sur des indices
library\ctlibstructureaddon\ct_itemdrawable\ct_grid4d_dense.h
    ==> Conversion de type implicite sur des indices
library\ctlibstructureaddon\ct_itemdrawable\ct_grid4d_sparse.h
    ==> Conversion de type implicite sur des indices
library\ctlibclouds\ct_cloudindex\ct_cloudindexstdlistt.hpp
    ==> Retour de fonction non intercepté
library\ctlibclouds\ct_cloudindex\tools\ct_cloudindexstdvectortmethodimpl.hpp
    ==> Retour de fonction non intercepté

git-svn-id: http://rdinnovation.onf.fr/svn/computreev6@95 b135c390-378b-4ba5-bdcd-e4540112eb5a

+++ AMKgl/tools/opengl/ellipseglrenderer.cpp
@@ -1,5 +1,3 @@
-#include "ellipseglrenderer.h"
-
 #if !defined (_USE_MATH_DEFINES)
 #define _USE_MATH_DEFINES
 #endif
@@ -7,6 +5,8 @@
 #include <QtMath>
 
 #include <QtOpenGL/QtOpenGL>
+
+#include "ellipseglrenderer.h"
 
 EllipseGLRenderer::EllipseGLRenderer()
 {
@@ -39,7 +39,7 @@ void EllipseGLRenderer::render()
 {
     const size_t& size = m_values.size();
 
-    for(int i=0; i<size; ++i)
+    for(int i=0; i<static_cast<int>(size); ++i)
         glVertex3fv(&m_values[i][0]);
 }
